Remove combine_keys method to its own Abstract Class
